refactor: :art: Moved formValidationService into /utils folder
